// src/shared/types/blocks/table.d.ts
import { Pagination } from './common';

export interface TableColumn {
  name?: string;
  title?: string;
  type?:
    | 'copy'
    | 'image'
    | 'time'
    | 'label'
    | 'dropdown'
    | 'user'
    | 'json_preview'
    | 'video';
  placeholder?: string;
  metadata?: any;
  className?: string;
  callback?: (item: any) => any;
}

export interface Table {
  title?: string;
  columns: TableColumn[];
  data: any[];
  emptyMessage?: string;
  pagination?: Pagination;
  actions?: Button[];
}